set desktop from path

+++ src/desktop.js
@@ -2,23 +2,25 @@ const imageSize = require('image-size');
 const wallpaper = require('wallpaper');
 
 
-module.exports = async function setDesktopImage(argFilenames) {
-  console.log('filtering images with minimum 1000px width requirement...\n');
-  const filenames = [];
-  const getPath = fname => `${__dirname}/../data/${fname}`
-  for (let filename of argFilenames) {
-    const pathToFile = getPath(filename);
+/**
+ *
+ * @param {string[]} argFilePaths array of paths to images to consider
+ * @param {boolean} shouldDeleteSmallImages if true will delete images whose width are less than 1000px
+ * @returns {string[]} paths to images left to delete if you opt out of caching the images
+ */
+module.exports = async function setDesktopImage(argFilePaths, shouldDeleteSmallImages = true) {
+  const filepaths = [];
+  for (let pathToFile of argFilePaths) {
     const { width } = imageSize(pathToFile);
-    if (width < 1e3) {
+    if (width < 1e3 && shouldDeleteSmallImages) {
       fs.unlinkSync(pathToFile);
     } else {
-      filenames.push(filename)
+      filepaths.push(pathToFile)
     }
   }
-  const index = Math.floor(Math.random() * filenames.length);
-  console.log(`${filenames[index]} selected for background.\n`)
-  const pathToFile = getPath(filenames[index]);
-  await wallpaper.set(pathToFile);
+  const index = Math.floor(Math.random() * filepaths.length);
+  console.log(`${filepaths[index]} selected for background.\n`);
+  await wallpaper.set(filepaths[index]);
   console.log('Background image set.\n');
-  return filenames;
+  return filepaths;
 }
